Relaxed the requirement regarding atomic override of destination registers for LD to allow more implementation options (#53)
* Relaxed the requirement regarding atomic override of destination registers for LD to allow more implementation options
The previous definition prevented some implementations for low end MCUs, which may be desirable.
Particularly, for devices with a 32b data interface and a 2R1W register file, atomic write back to the pair would only be possible by disabling interrupts during the operation.
The requirement for atomic writeback was unnecessarily restrictive.
This change relaxed the requirement to the minimum needed to ensure fault handling is possible.
The change is backward compatible as any implementation complying to the more restrictive requirement complies to the updated one

---------

Signed-off-by: Christian Herber <christian.herber@oss.nxp.com>